requirements_gemini.txt
# Google Gemini API Integration Requirements
# Install with: pip install -r requirements_gemini.txt

# Google Generative AI SDK (Gemini API)
google-genai>=1.0.0

# Google Cloud Speech-to-Text (if using Google STT)
google-cloud-speech>=2.20.0

# Text-to-Speech engines
gTTS>=2.4.0
edge-tts>=6.1.0
pydub>=0.25.1

# Core dependencies (keep from original requirements.txt)
pyyaml>=6.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0
colorama>=0.4.6
psutil>=5.9.0
diskcache>=5.6.0
SpeechRecognition>=3.10.0

# MCP Server
mcp>=0.1.0

# Audio Recording for Speech Chat
streamlit-audiorec>=0.1.3

# WebRTC for Live Voice Chat
streamlit-webrtc>=0.64.0
aiortc>=1.14.0
av>=10.0.0
